+++ tests/test_plugin.py
@@ -1,0 +1,13 @@
+from pytest_helm_charts.fixtures import _parse_extra_info  # noqa
+
+
+def test_parse_chart_extra_info():
+    res = _parse_extra_info("key1=val1,external_cluster_type=kind")
+    assert len(res) == 2
+    assert "key1" in res and res["key1"] == "val1"
+    assert "external_cluster_type" in res and res["external_cluster_type"] == "kind"
+
+
+def test_parse_chart_no_extra_info():
+    res = _parse_extra_info("")
+    assert len(res) == 0