+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), following [Se
 
 ## [Unreleased]
 
+## [0.1.5]
+- add: `--chart-extra-info` CLI option that can be used to pass arbitrary `key=value` maps to the test cases.
+  Values fo the map are accessible in code as `chart_extra_info` fixture.
+
 ## [0.1.4]
 - fix: wait_for_namespaced_objects_conditions was exiting prematurely when the objects were not yet present